Do not call git for the user
IMO the user is responsible for Git, and we shouldn't try to run git
commands (even though they are completely safe).
Futhermore the `git add` one is not so useful, since the user still has
to modify the check before it's usable.
I replaced the `git rm` by a suggestion to run git rm.