[ci] use requirements file to install pip deps
It's way faster than iterating over all the dependencies.
The only drawback is that if one dep fails to install, the other won't
be installed.
This was an issue in `dd-agent` as all the requirements for the check
were in one file, meaning that to be able to run the tests for haproxy
you had to try to install psycopg2 (requirement for the pgbouncer check).
This is not the case anymore! 🎉🎉🎉🎉🎉